+++ test/ForkIOScheduler.hs
@@ -1,5 +1,10 @@
 module ForkIOScheduler where
 
+import Control.Exception
+import Control.Concurrent
+import Control.Concurrent.STM
+import Control.Eff
+import Control.Eff.Lift
 import Control.Eff.Concurrent.Process
 import Control.Eff.Concurrent.Process.ForkIOScheduler as Scheduler
 import Control.Monad (void, forever)
@@ -9,6 +14,41 @@ import Data.Dynamic
 
 timeoutSeconds :: Integer -> Timeout
 timeoutSeconds seconds = Timeout (seconds * 1000000) (show seconds ++ "s")
+
+test_IOExceptionsIsolated :: TestTree
+test_IOExceptionsIsolated =
+  localOption
+  (timeoutSeconds 30)
+    $ testGroup "one process throws an IO exception, the other continues unimpaired"
+       [ testCase ("process 2 exits with: "++ howToExit
+                    ++ " - while process 1 is busy with: " ++ busyWith)
+        $ do aVar <- newEmptyTMVarIO
+             Scheduler.defaultMain
+               $ do p1 <- spawn $ forever busyEffect
+                    lift (threadDelay 1000)
+                    void $ spawn $ do lift (threadDelay 1000)
+                                      doExit
+                    lift (threadDelay 100000)
+                    wasStillRunningP1 <- sendShutdown forkIoScheduler p1
+                    lift (atomically (putTMVar aVar wasStillRunningP1))
+
+             wasStillRunningP1 <- atomically (takeTMVar aVar)
+             assertBool "the other process was still running" wasStillRunningP1
+
+       | (busyWith, busyEffect) <-
+         [ ("receiving", void (send (ReceiveMessage @SchedulerIO)))
+         , ("sending", void (send (SendMessage @SchedulerIO 44444 (toDyn "test message"))))
+         , ("sending shutdown", void (send (SendShutdown @SchedulerIO 44444)))
+         , ("selfpid-ing", void (send (SelfPid @SchedulerIO)))
+         , ("spawn-ing", void (send (Spawn @SchedulerIO (void (send (ReceiveMessage @SchedulerIO))))))
+         ]
+       , (howToExit, doExit) <-
+         [ ("throw async exception", void (lift (throw UserInterrupt)))
+         , ("division by zero", void (return ((123::Int) `div` 0) >>= lift . putStrLn . show ))
+         , ("call 'fail'", void (fail "test fail"))
+         , ("call 'error'", void (error "test error"))
+         ]
+       ]
 
 test_mainProcessSpawnsAChildAndReturns :: TestTree
 test_mainProcessSpawnsAChildAndReturns =